+++ threedi_models_and_simulations/data_models/simulation_data_models.py
@@ -74,7 +74,7 @@ class InitOptions(SimulationElement):
 @dataclass
 class BoundaryConditions(SimulationElement):
     file_boundary_conditions: FileBoundaryCondition = None
-    data: dict = None
+    data: list = None
 
 
 @dataclass